Added null checks to local storage functions

+++ utils/browserHandle.ts
@@ -4,19 +4,19 @@
 import { DataGridSettingsValues } from "../types";
 
 export function getLocalStorageItem(key: string) {
-    return localStorage.getItem(key);
+    return typeof window !== 'undefined' ? localStorage.getItem(key) : "";
 }
 
 export function setLocalStorageItem(key: string, value: string) {
-    localStorage.setItem(key, value);
+    if (typeof window !== 'undefined') localStorage.setItem(key, value);
 }
 
 export function removeLocalStorageItem(key: string) {
-    localStorage.removeItem(key);
+    if (typeof window !== 'undefined') localStorage.removeItem(key);
 }
 
 export function clearLocalStorage() {
-    localStorage.clear();
+    if (typeof window !== 'undefined') localStorage.clear();
 }
 
 export function getLocalStorageKeys() {
